techpack: audio: DSP: Force V3 functions for legacy ADSP
The drivers are sometimes calling the V4 functions directly for
some monkey reason...
This makes the DSP to go completely nuts on older platforms,
since the firmware is not updated to support the V4 functions.

To address this, even if - okay, it's ugly - add ifdefs to force
calling the corresponding V3 functions: this makes sure that we
won't get breakage from random CAF tags upgrades and such.

Change-Id: Ia8432854364d9f0df443f3aab43d35a93070e4ed