ARM64: dts: msm: disable memcg kernel and socket accounting on msm8956/76
Disable memcg kernel and socket accounting on sdm845 since
memory cgroups is used in this target as a means of user space
task grouping based on oom_score_adj, and controlling the userspace
memory consumed by tasks. Kernel and socket memory accounting is
not considered. Since memcg is created per process on this target
enabling kernel accounting results in the creation of numerous
kmem caches resulting in significant overhead.

Change-Id: I2caad9ce6cca5846b6183a0f0753977db54a660d
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>

Change-Id: I2d204fb361ce47b03d8b58ce021a67749eeb25f6
Signed-off-by: Ricky Cheung <rcheung844@gmail.com>