power: oppo: Add initial VOOC drivers
VOOC is a fast charging technology developed by OPPO Electronics
which was introduced in 2014. Early versions of this tech utilizes
a pic16f chip for communication of charging status. Some versions
also used a stm8s chip.

This driver implements basic communication with the chips, requiring
constant polling of information to check whether a valid adapter has
been connected, and then switching GPIOs to enable charging.

FW update support needs to be added for full functionality. In the
downstream ports provided by OPPO, they had included the firmware
inside of the kernel, this has been discouraged since quite a few
years ago upstream, so the implementation still needs to be
confirmed for sanity.

Change-Id: Ie039298fe70813d41abe18ce22936204326d4417
Signed-off-by: Ricky Cheung <rcheung844@gmail.com>